fitness_tracker/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
pytest==7.4.2
SQLAlchemy==2.1.0
